// ui/mill-grinder-ui/src/api/mill/api.ts
export interface TreeNodeDto {
    /**
     * 
     * @type {string}
     * @memberof TreeNodeDto
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof TreeNodeDto
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof TreeNodeDto
     */
    'type'?: TreeNodeDtoTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof TreeNodeDto
     */
    'displayName'?: string;
    /**
     * 
     * @type {string}
     * @memberof TreeNodeDto
     */
    'description'?: string;
    /**
     * 
     * @type {Array<TreeNodeDto>}
     * @memberof TreeNodeDto
     */
    'children'?: Array<TreeNodeDto>;
    /**
     * 
     * @type {boolean}
     * @memberof TreeNodeDto
     */
    'hasChildren'?: boolean;
}

export const TreeNodeDtoTypeEnum = {
    Catalog: 'CATALOG',
    Schema: 'SCHEMA',
    Table: 'TABLE',
    Attribute: 'ATTRIBUTE',
    Concept: 'CONCEPT'
} as const;

export type TreeNodeDtoTypeEnum = typeof TreeNodeDtoTypeEnum[keyof typeof TreeNodeDtoTypeEnum];